refactor(score): separate selection bridge and tests
